helper/requirements.txt
google-genai
internetarchive
moviepy
notebook
pandas
pymongo
python-dotenv
